% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areapiabusummary.R
\name{abupicreate}
\alias{abupicreate}
\title{abupicreate}
\usage{
abupicreate(habitatcreationdata, sectioncounter)
}
\arguments{
\item{habitatcreationdata}{hab loss df}

\item{sectioncounter}{counter passed to following dynamic area based post intervention sections}
}
\value{
sectioncounter, passed to future dynamically rendered sections under post intervention
}
\description{
abupicreate
}
\examples{
\dontrun{sectioncounter<-abupicreate(habitatcreationdata$habitatcreationdata, sectioncounter)}
}
