% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areapiabusummary.R
\name{abupienhance}
\alias{abupienhance}
\title{abupienhance}
\usage{
abupienhance(habitatenhancementdata, sectioncounter)
}
\arguments{
\item{habitatenhancementdata}{hab en df}

\item{sectioncounter}{counter passed to following dynamic area based post intervention sections}
}
\value{
sectioncounter, passed to future dynamically rendered sections under post intervention
}
\description{
abupienhance
}
\examples{
\dontrun{sectioncounter<- abupienhance(habitatenhancedata$habitatenhancementdata, sectioncounter)}
}
