% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areapiabusummary.R
\name{abupilost}
\alias{abupilost}
\title{abupilost}
\usage{
abupilost(habitatlostdata, sectioncounter)
}
\arguments{
\item{habitatlostdata}{hab loss df}

\item{sectioncounter}{counter passed to following dynamic area based post intervention sections}
}
\value{
sectioncounter, passed to future dynamically rendered sections under post intervention
}
\description{
abupilost
}
\examples{
\dontrun{sectioncounter<-abupilost(habitatlossdata$habitatlostdata, sectioncounter}
}
