% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areapiabusummary.R
\name{abupiretain}
\alias{abupiretain}
\title{abupiretain}
\usage{
abupiretain(habitatretaindata, sectioncounter)
}
\arguments{
\item{habitatretaindata}{hab loss df}

\item{sectioncounter}{counter passed to following dynamic area based post intervention sections}
}
\value{
sectioncounter, passed to future dynamically rendered sections under post intervention
}
\description{
abupiretain
}
\examples{
\dontrun{sectioncounter<-abupiretain(habitatretaindata$habitatretaindata, sectioncounter)}
}
