% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abusummary.R
\name{abusummary}
\alias{abusummary}
\title{abusummary}
\usage{
abusummary(habitatbaselinedata, sectioncounter)
}
\arguments{
\item{habitatbaselinedata}{baselinehabdata}

\item{sectioncounter}{counter to render numbers}
}
\value{
printed text
}
\description{
abusummary
}
\examples{
\dontrun{abusummary(habitatbaselinedata, sectioncounter)}
}
