% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data_catches.R
\name{check_A1}
\alias{check_A1}
\title{cleans onsitehab baseline metric data}
\usage{
check_A1(metric)
}
\arguments{
\item{metric}{a filepath to a metric}
}
\value{
error messages, a string of error messages to be shown so user knows
where data is incorrect
}
\description{
cleans onsitehab baseline metric data
}
\examples{
\dontrun{checked <- check_A1(metric = system.file("extdata",
 "OnSiteBoth.xlsx", package = "metricpulling"))}
}
