% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data_catches.R
\name{check_A3}
\alias{check_A3}
\title{cleans onsitehab enchancement metric data}
\usage{
check_A3(metric)
}
\arguments{
\item{metric}{feas metric}
}
\value{
df, a cleaned version of the input dataset - to be used as an argument in actual metric pulling function
}
\description{
cleans onsitehab enchancement metric data
}
\examples{
\dontrun{checked_dataset<-check_A3(metric)}   
}
