% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data_catches.R
\name{check_B2}
\alias{check_B2}
\title{cleans onsitehedge creation loss data}
\usage{
check_B2(metric)
}
\arguments{
\item{metric}{feas metric}
}
\value{
df, a cleaned version of the input dataset - to be used as an argument in actual metric pulling function
}
\description{
cleans onsitehedge creation loss data
}
\examples{
\dontrun{checked_dataset<-check_B2(metric)}    
}
