% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data_catches.R
\name{check_B3}
\alias{check_B3}
\title{cleans onsitehedge enhancement loss data}
\usage{
check_B3(metric)
}
\arguments{
\item{metric}{feas metric}
}
\value{
df, a cleaned version of the input dataset - to be used as an argument in actual metric pulling function
}
\description{
cleans onsitehedge enhancement loss data
}
\examples{
\dontrun{checked_dataset<-check_B3(metric)}   
}
