% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data_catches.R
\name{check_C1}
\alias{check_C1}
\title{cleans C1 water metric data}
\usage{
check_C1(metric)
}
\arguments{
\item{metric}{feas metric}
}
\value{
df, a cleaned version of the input dataset - to be used as an argument in actual metric pulling function
}
\description{
cleans C1 water metric data
}
\examples{
\dontrun{checked_dataset<-check_C1(metric)}  
}
