% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data_catches.R
\name{checkhabsum_dataset}
\alias{checkhabsum_dataset}
\title{cleans habsum metric data}
\usage{
checkhabsum_dataset(metric)
}
\arguments{
\item{metric}{feas metric}
}
\value{
df, a cleaned version of the input dataset - to be used as an argument in actual metric pulling function
}
\description{
cleans habsum metric data
}
\examples{
\dontrun{checked_dataset<-checkhabsum_dataset(metric)}   
}
