% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data_catches.R
\name{checkhedgesum_dataset}
\alias{checkhedgesum_dataset}
\title{cleans hedgesum metric data}
\usage{
checkhedgesum_dataset(metric)
}
\arguments{
\item{metric}{feas metric}
}
\value{
df, a cleaned version of the input dataset - to be used as an argument in actual metric pulling function
}
\description{
cleans hedgesum metric data
}
\examples{
\dontrun{checked_dataset<-checkhedgesum_dataset(metric)}     
}
