% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data_catches.R
\name{checkonsitenet_dataset}
\alias{checkonsitenet_dataset}
\title{cleans onsitenet metric data}
\usage{
checkonsitenet_dataset(metric)
}
\arguments{
\item{metric}{feas metric}
}
\value{
df, a cleaned version of the input dataset - to be used as an argument in actual metric pulling function
}
\description{
cleans onsitenet metric data
}
\examples{
\dontrun{checked_dataset<-checkonsitenet_dataset(metric)}    
}
