% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data_catches.R
\name{checkwatersum_dataset}
\alias{checkwatersum_dataset}
\title{cleans watersum metric data}
\usage{
checkwatersum_dataset(metric)
}
\arguments{
\item{metric}{feas metric}
}
\value{
df, a cleaned version of the input dataset - to be used as an argument in actual metric pulling function
}
\description{
cleans watersum metric data
}
\examples{
\dontrun{checked_dataset<-checkwatersum_dataset(metric)}
}
