% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data_catches.R
\name{clean_onsitehab_baseline}
\alias{clean_onsitehab_baseline}
\title{cleans onsitehab baseline metric data}
\usage{
clean_onsitehab_baseline(metric)
}
\arguments{
\item{metric}{feas metric}
}
\value{
error messages, a string of error messages to be shown so user knows where data is incorrect
}
\description{
cleans onsitehab baseline metric data
}
\examples{
\dontrun{checked_dataset<-clean_onsitehab_baseline(metric)}   
}
