% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data_catches.R
\name{clean_onsitehab_creation}
\alias{clean_onsitehab_creation}
\title{cleans onsitehab creation metric data}
\usage{
clean_onsitehab_creation(metric)
}
\arguments{
\item{metric}{feas metric}
}
\value{
df, a cleaned version of the input dataset - to be used as an argument in actual metric pulling function
}
\description{
cleans onsitehab creation metric data
}
\examples{
\dontrun{checked_dataset<-clean_onsitehab_creation(metric)}     
}
