% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data_catches.R
\name{clean_onsitehab_loss}
\alias{clean_onsitehab_loss}
\title{cleans onsitehab loss metric data}
\usage{
clean_onsitehab_loss(metric)
}
\arguments{
\item{metric}{feas metric}
}
\value{
df, a cleaned version of the input dataset - to be used as an argument in actual metric pulling function
}
\description{
cleans onsitehab loss metric data
}
\examples{
\dontrun{checked_dataset<-clean_onsitehab_loss(metric)}       
}
