% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data_catches.R
\name{clean_onsitehedge_baseline}
\alias{clean_onsitehedge_baseline}
\title{cleans onsitehedge baseline metric data}
\usage{
clean_onsitehedge_baseline(metric)
}
\arguments{
\item{metric}{feas metric}
}
\value{
df, a cleaned version of the input dataset - to be used as an argument in actual metric pulling function
}
\description{
cleans onsitehedge baseline metric data
}
\examples{
\dontrun{checked_dataset<-clean_onsitehedge_baseline(metric)}     
}
