% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data_catches.R
\name{clean_onsitehedge_loss}
\alias{clean_onsitehedge_loss}
\title{cleans onsitehedge retain loss data}
\usage{
clean_onsitehedge_loss(metric)
}
\arguments{
\item{metric}{feas metric}
}
\value{
df, a cleaned version of the input dataset - to be used as an argument in actual metric pulling function
}
\description{
cleans onsitehedge retain loss data
}
\examples{
\dontrun{checked_dataset<-clean_onsitehedge_loss(metric)}    
}
