% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data_catches.R
\name{clean_onsitenet_dataset}
\alias{clean_onsitenet_dataset}
\title{cleans onsitenet metric data}
\usage{
clean_onsitenet_dataset(metric)
}
\arguments{
\item{metric}{feas metric}
}
\value{
df, a cleaned version of the input dataset - to be used as an argument in actual metric pulling function
}
\description{
cleans onsitenet metric data
}
\examples{
\dontrun{checked_dataset<-clean_onsitenet_dataset(metric)}    
}
