% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/developmentstrings.R
\name{developmentstring}
\alias{developmentstring}
\title{dev string}
\usage{
developmentstring(losthabitattype, losthedgetype, lostwatertype)
}
\arguments{
\item{losthabitattype}{lost habs}

\item{losthedgetype}{lost hedges}

\item{lostwatertype}{lost waters}
}
\value{
string for development
}
\description{
dev string
}
\examples{
\dontrun{developmentstring<-developmentstring(habitatlossdata$habitatlostdata$habitattype, hedgelossdata$hedgelostdata$habitattype, waterlossdata$waterlostdata$habitattype)}
}
