% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executivesummarystrings.R
\name{exsum_baselinestring}
\alias{exsum_baselinestring}
\title{exsum baseline string}
\usage{
exsum_baselinestring(habitattype)
}
\arguments{
\item{habitattype}{the detailed habitat type, e.g. 'Other neutral grassland', contained in the list
produced by the function 'pulloffsitebaseline'.}
}
\value{
the strings of the unique habitat types present on site
}
\description{
generates the summary of the unique baseline habitats on site, to be printed in the executive summary
}
\examples{
\dontrun{baselinestring<-exsum_baselinestring(habitattype)}

}
