% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executivesummarystrings.R
\name{exsum_creationstring}
\alias{exsum_creationstring}
\title{exsum creation string}
\usage{
exsum_creationstring(createdhabitattype)
}
\arguments{
\item{createdhabitattype}{the detailed habitat type, e.g. 'Other neutral grassland', contained in the list
produced by the function 'pulloffsitecreation'.}
}
\value{
the strings of the unique habitat types to be created, returning 'nothing' if none
}
\description{
generates the summary of the unique habitats to be created on site, to be printed in the executive summary
}
\examples{
\dontrun{creationstring<-exsum_creationstring(creationdata$createdhabitattype)}
}
