% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executivesummarystrings.R
\name{exsum_enhancestring}
\alias{exsum_enhancestring}
\title{exsum enhancement string}
\usage{
exsum_enhancestring(enhancedhabitattype)
}
\arguments{
\item{enhancedhabitattype}{the detailed habitat type, e.g. 'Other neutral grassland', contained in the list
produced by the function 'pulloffsiteenhancement'.}
}
\value{
the strings of the unique habitat types to be enhanced, returning 'nothing' if none
}
\description{
exsum enhancement string
}
\examples{
\dontrun{enhancestring<- exsum_enhancestring(enhancedata$enhancedhabitattype)}
}
