% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executivesummarystrings.R
\name{exsum_loststring}
\alias{exsum_loststring}
\title{exsum loss string}
\usage{
exsum_loststring(losthabitattype)
}
\arguments{
\item{losthabitattype}{the detailed habitat type, e.g. 'Other neutral grassland', contained in the list
produced by the function 'pulloffsitehabitatbaseline'.}
}
\value{
the strings of the unique habitat types to be lost, returning 'nothing' if none
}
\description{
exsum loss string
}
\examples{
\dontrun{loststring<- exsum_loststring(baselinedata$losthabitattype)}
}
