% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executivesummarystrings.R
\name{exsum_passstring}
\alias{exsum_passstring}
\title{passstring}
\usage{
exsum_passstring(netdata)
}
\arguments{
\item{netdata}{data showing if all linear and area based units have reached net gain, and if trading summaries are met}
}
\value{
a string reporting if the proposals passed, and if not, what theyre mmissing
}
\description{
passstring
}
\examples{
\dontrun{passstring<-exsum_passstring(netdata)}
}
