% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposallist.R
\name{feas_proposedlist}
\alias{feas_proposedlist}
\title{proposed list, to be used in X section}
\usage{
feas_proposedlist(enhancedata, creationdata)
}
\arguments{
\item{enhancedata}{produced by pulloffsiteenhancement}

\item{creationdata}{produced by pulloffsitecreation}
}
\value{
string of proposed habitat types to be used in proposal section
}
\description{
proposed list, to be used in X section
}
\examples{
\dontrun{proposedlist<-feas_proposedlist(enhancedata, creationdata)}
}
