% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hedgelbupi.R
\name{hedgelbupicreate}
\alias{hedgelbupicreate}
\title{hedgelbupicreate}
\usage{
hedgelbupicreate(hedgecreationdata, sectioncounter)
}
\arguments{
\item{hedgecreationdata}{hedge creation data}

\item{sectioncounter}{counter passed to following dynamic area based post intervention sections}
}
\value{
sectioncounter, passed to future dynamically rendered sections under post intervention
}
\description{
hedgelbupicreate
}
\examples{
\dontrun{sectioncounter<-hedgelbupicreate(hedgecreationdata$hedgecreationdata, sectioncounter)}
}
