% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hedgelbupi.R
\name{hedgelbupienhance}
\alias{hedgelbupienhance}
\title{linear enhance}
\usage{
hedgelbupienhance(hedgeenhancementdata, sectioncounter)
}
\arguments{
\item{hedgeenhancementdata}{hedge df}

\item{sectioncounter}{counter to pass}
}
\value{
counter to pass
}
\description{
linear enhance
}
\examples{
\dontrun{sectioncounter<-hedgelbupienhance(hedgeenhancedata$hedgeenhancementdata, sectioncounter)}
}
