% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hedgelbupi.R
\name{hedgelbupilost}
\alias{hedgelbupilost}
\title{hedgelbupilost}
\usage{
hedgelbupilost(hedgelostdata, sectioncounter)
}
\arguments{
\item{hedgelostdata}{hedge loss df}

\item{sectioncounter}{counter passed to following dynamic area based post intervention sections}
}
\value{
sectioncounter, passed to future dynamically rendered sections under post intervention
}
\description{
hedgelbupilost
}
\examples{
\dontrun{sectioncounter <- hedgelbupilost(hedgelossdata$hedgelostdata, sectioncounter)}
}
