% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hedgelbupi.R
\name{hedgelbupiretain}
\alias{hedgelbupiretain}
\title{hedgelbupiretain}
\usage{
hedgelbupiretain(hedgeretaindata, sectioncounter)
}
\arguments{
\item{hedgeretaindata}{hedge retain df}

\item{sectioncounter}{counter passed to following dynamic area based post intervention sections}
}
\value{
sectioncounter, passed to future dynamically rendered sections under post intervention
}
\description{
hedgelbupiretain
}
\examples{
#' @examples \dontrun{sectioncounter<-hedgelbupiretain(hedgeretaindata$hedgeretaindata, sectioncounter)}
}
