% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/introductionstrings.R
\name{intro_habbaselinelist}
\alias{intro_habbaselinelist}
\title{baseline list, the list of baseline habitats on site}
\usage{
intro_habbaselinelist(baselinedata)
}
\arguments{
\item{baselinedata}{output of pulloffsitebaseline}
}
\value{
list of strings containing the baseline habitats on site
}
\description{
baseline list, the list of baseline habitats on site
}
\examples{
\dontrun{baselinehabitatlist<-intro_habbaselinelist(baselinedata)}
}
