% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/introductionstrings.R
\name{intro_linearbaselinelist}
\alias{intro_linearbaselinelist}
\title{baseline list, the list of baseline habitats on site}
\usage{
intro_linearbaselinelist(baselinedata)
}
\arguments{
\item{baselinedata}{output of pulloffsitebaseline}
}
\value{
list of strings containing the baseline habitats on site
}
\description{
baseline list, the list of baseline habitats on site
}
\examples{
\dontrun{baselinewaterlist<-intro_linearbaselinelist(baselinedata)}
}
