% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/introductionstrings.R
\name{lpastring}
\alias{lpastring}
\title{get lpa string for introduction section}
\usage{
lpastring(lpa_names)
}
\arguments{
\item{lpa_names}{the dataframe pulled by getlpa, containing the name(s) of
the LPA district the site is in.}
}
\value{
a string containing the LPAs the site is in.
}
\description{
get lpa string for introduction section
}
\examples{
\dontrun{lpastring<-lpastring(lpa_names)}
}
