% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_check.R
\name{metric_check}
\alias{metric_check}
\title{metric_check}
\usage{
metric_check(metric)
}
\arguments{
\item{metric}{a file path to a metric}
}
\value{
text with list of where data is missing in metric
}
\description{
metric_check
}
\examples{
metric_check(system.file("extdata", "OnSiteBoth.xlsx", package = "metricpulling"))
}
