% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricpulling-package.R
\docType{package}
\name{metricpulling-package}
\alias{metricpulling}
\alias{metricpulling-package}
\title{metricpulling: Pulling data from DEFRA BNG Metric}
\description{
Pulling data from the DEFRA Biodiversity Net Gain Metric.
}
\author{
\strong{Maintainer}: Dan Carpenter \email{info@digital-ecology.co.uk}

Authors:
\itemize{
  \item Maddie Lucas
  \item Charlie Hewitt
}

}
\keyword{internal}
