% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newbaselinefunction.R
\name{newonsitehabitatbaseline}
\alias{newonsitehabitatbaseline}
\title{New on site habitat baseline function for testing}
\usage{
newonsitehabitatbaseline(metric)
}
\arguments{
\item{metric}{file path to metric}
}
\value{
list with dataframe of baseline data, total area nd total baselien units
}
\description{
New on site habitat baseline function for testing
}
\examples{
\dontrun{baselinedata<-newonsitehabitatbaseline(metric)}
}
