% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newbaselinefunction.R
\name{newonsitehabitatcreation}
\alias{newonsitehabitatcreation}
\title{New on site habitat created function for testing}
\usage{
newonsitehabitatcreation(metric)
}
\arguments{
\item{metric}{fiel path to metric}
}
\value{
list with data frame of habitats created, total area of habitats created, and total units created
}
\description{
New on site habitat created function for testing
}
\examples{
\dontrun{newonsitehabitatcreation(metric)}
}
