% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newbaselinefunction.R
\name{newonsitehabitatloss}
\alias{newonsitehabitatloss}
\title{New on site habitat lost function for testing}
\usage{
newonsitehabitatloss(metric)
}
\arguments{
\item{metric}{file path to metric}
}
\value{
list with dataframe of habitats lost, total area lost, and total units lost
}
\description{
New on site habitat lost function for testing
}
\examples{
\dontrun{newonsitehabitatloss(metric)}
}
