% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newbaselinefunction.R
\name{newonsitehabitatretain}
\alias{newonsitehabitatretain}
\title{New on site habitats retained function for testing}
\usage{
newonsitehabitatretain(metric)
}
\arguments{
\item{metric}{file path to metric}
}
\value{
list with dataframe of retained habitats, total retained area, total retained units
}
\description{
New on site habitats retained function for testing
}
\examples{
\dontrun{newonsitehabitatretain(metric)}
}
