% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offsite_habitat_metricfunctions.R
\name{offsitehabitatbaseline}
\alias{offsitehabitatbaseline}
\title{Offsite baseline habitat data for summary}
\usage{
offsitehabitatbaseline(metric)
}
\arguments{
\item{metric}{BNG metric}
}
\value{
a list, containing dataframes pulled from the offfsite baseline sheet of the metric
}
\description{
Offsite baseline habitat data for summary
}
\examples{
\dontrun{baselinedata<-offsitehabitatbaseline(metric)}
}
