% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offsite_habitat_metricfunctions.R
\name{offsitehabitatcreation}
\alias{offsitehabitatcreation}
\title{Offsite habitat creation data summary}
\usage{
offsitehabitatcreation(metric)
}
\arguments{
\item{metric}{A BNG metric}
}
\value{
A list with a summary data.frame, an numeric for total habitat area
created and a numeric for total BNG units created
}
\description{
Offsite habitat creation data summary
}
\examples{
\dontrun{creationdata <- offsitehabitatcreation(metric)}
}
