% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offsite_habitat_metricfunctions.R
\name{offsitehabitatenhance}
\alias{offsitehabitatenhance}
\title{Off-site habitat enhancement summary data}
\usage{
offsitehabitatenhance(metric)
}
\arguments{
\item{metric}{A BNG metric}
}
\value{
A list with a summary data.frame, an numeric for total habitat area
enhanced and a numeric for total BNG units enhanced
}
\description{
Off-site habitat enhancement summary data
}
\examples{
\dontrun{enhancedata <- offsitehabitatenhance(metric)}
}
