% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offsite_habitat_metricfunctions.R
\name{offsitehabitatgainsummary}
\alias{offsitehabitatgainsummary}
\title{Off-site habitat gain summary}
\usage{
offsitehabitatgainsummary(metric)
}
\arguments{
\item{metric}{A BNG metric}
}
\value{
A data.frame of summary data for off-site biodiversity net gains
}
\description{
Off-site habitat gain summary
}
\examples{
\dontrun{netgainsummary <- offsitehabitatgainsummary(metric)}
}
