% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offsite_habitat_metricfunctions.R
\name{offsitehabitatloss}
\alias{offsitehabitatloss}
\title{Off site habitats lost data}
\usage{
offsitehabitatloss(metric)
}
\arguments{
\item{metric}{A BNG metric}
}
\value{
A list with a summary data.frame, an numeric for total habitat area
lost and a numeric for total BNG units lost
}
\description{
Off site habitats lost data
}
\examples{
\dontrun{lossdata<-offsitehabitatloss(metric)}
}
