% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offsite_habitat_metricfunctions.R
\name{offsitehabitatretain}
\alias{offsitehabitatretain}
\title{Offsite retained habitat data for summary}
\usage{
offsitehabitatretain(metric)
}
\arguments{
\item{metric}{A BNG metric}
}
\value{
A list with a summary data.frame, an numeric for total area retained
and a numeric for total BNG units retained
}
\description{
Offsite retained habitat data for summary
}
\examples{
\dontrun{retaineddata<-offsitehabitatretain(metric)}
}
