% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OnSiteHedgeEnhance}
\alias{OnSiteHedgeEnhance}
\title{On-Site Hedge Enhance Data}
\format{
A list containing multiple data frames, where each data frame corresponds to
a sheet in the original Excel file.

The dataset consists of the following sheets:
\describe{
\item{Introduction}{General overview of the dataset and methodology.}
\item{Start}{Initial setup and user inputs.}
\item{Main Menu}{Navigation menu for the calculations.}
\item{Unit shortfall summary}{Summary of biodiversity unit shortfalls.}
\item{Results}{Aggregated results of biodiversity calculations.}
\item{Headline Results}{Summary of biodiversity unit shortfalls.}
\item{Detailed Results}{Aggregated results of biodiversity calculations.}
\item{Trading Summary Area Habitats}{Summary of trading values for habitats.}
\item{Trading Summary Hedgerows}{Summary of trading values for hedgerows.}
\item{Trading Summary WaterC's}{Summary of trading values for watercourses.}
\item{Off-site gain site summary}{Summary of off-site unit gains for habitat, hedgerow and waterways.}
\item{Irreplaceable Habitats}{Summary of irreplaceable habitats on and off-site.}
\item{A-1 On-Site Habitat Baseline}{Summary of trading values for on-site habitats}
\item{A-2 On-Site Habitat Creation}{Planned habitat creation for on-site locations.}
\item{A-3 On-Site Habitat Enhancement}{Planned habitat enhancements for on-site locations.}
\item{D-1 Off-Site Habitat Baseline}{Baseline condition of off-site habitats.}
\item{D-2 Off-Site Habitat Creation}{Planned habitat creation for off-site locations.}
\item{D-3 Off-Site Habitat Enhancement}{Planned habitat enhancements for off-site locations.}
\item{B-1 On-Site Hedge Baseline}{Baseline condition of on-site hedges}
\item{B-2 On-Site Hedge Creation}{Planned hedges creation for on-site locations.}
\item{B-3 On-Site Hedge Enhancement}{Planned hedges enhancements for on-site locations.}
\item{E-1 Off-Site Hedge Baseline}{Baseline condition of off-site hedges}
\item{E-2 Off-Site Hedge Creation}{Planned hedges creation for off-site locations.}
\item{E-3 Off-Site Hedge Enhancement}{Planned habitat enhancements for off-site hedges}
\item{C-1 On-Site WaterC' Baseline}{Baseline condition of on-site water}
\item{C-2 On-Site WaterC' Creation}{Planned water creation for on-site locations.}
\item{C-3 On-Site WaterC' Enhancement}{Planned water enhancements for on-site locations.}
\item{F-1 Off-Site WaterC' Baseline}{Baseline condition of off-site waterways}
\item{F-2 Off-Site WaterC' Creation}{Planned water creation for off-site locations.}
\item{F-3 Off-Site WaterC' Enhancement}{Planned water enhancements for off-site locations.}
\item{Unit shortfall calculations}{Unit shortfall calculations for all habitat types.}
\item{G-1 All Habitats}{A hierarchical classification of habitats based on UKHab, including distinctiveness scores, trading rules, and technical difficulty for creation and enhancement.}
\item{G-2 Habtiat groups}{Summary of habitat areas, units, and net changes on-site and off-site before and after development.}
\item{G-3 Multipliers}{Comprehensive list of risk and multipliers for each habitat type.}
\item{G-4 Temporal multipliers}{Comprehensive list of temporal multipliers for each habitat type}
\item{G-5 Enhancement temporal}{Comprehensive list of start and target condition for all habitats used in the dataset.}
\item{G-6 Hedgerow Data}{Details on distinctiveness, target condition timelines, difficulty, trading rules, and post-development hedge type.}
\item{G-7 WaterC' Data}{Distinctiveness, difficulty, and condition for each water habtiat type.}
\item{G-8 Condition Look up}{Summary of conditions for each habitat type.}
\item{Version History}{Summary of changes made, versions and dates.}
\item{Phase 1 Translation Tool}{Comprehensive list of all habitats and their distintiveness band.}
\item{Technical Data}{Reference data for calculations and methodologies.}
\item{Lists}{Additional lookup tables and reference data.}
}
}
\source{
Biodiversity Net Gain calculations
}
\usage{
OnSiteHedgeEnhance
}
\description{
This dataset contains multiple sheets related to habitat creation, trading summaries,
and biodiversity net gain calculations for on-site habitats.It includes hedge and waterways data.
}
\keyword{datasets}
