% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OnSiteTradingSummary}
\alias{OnSiteTradingSummary}
\title{On-Site Trading Summmary Data}
\format{
A list containing multiple data frames, where each data frame corresponds to
a sheet in the original Excel file.

The dataset consists of the following sheets:
\describe{
\item{Start}{Initial setup and user inputs.}
\item{Main Menu}{Navigation menu for the calculations.}
\item{Results}{Aggregated results of biodiversity calculations.}
\item{Headline Results}{Summary of biodiversity unit shortfalls.}
\item{Trading Summary Area Habitats}{Summary of trading values for habitats.}
\item{Trading Summary Hedgerows}{Summary of trading values for hedgerows.}
\item{Trading Summary WaterC's}{Summary of trading values for watercourses.}
\item{A-1 On-Site Habitat Baseline}{Summary of trading values for on-site habitats}
\item{A-2 On-Site Habitat Creation}{Planned habitat creation for on-site locations.}
\item{B-1 On-Site Hedge Baseline}{Baseline condition of on-site hedges}
\item{B-2 On-Site Hedge Creation}{Planned hedges creation for on-site locations.}
}
}
\source{
Biodiversity Net Gain calculations
}
\usage{
OnSiteTradingSummary
}
\description{
This dataset contains multiple sheets related to habitat creation, trading summaries,
and biodiversity net gain calculations for on-site habitats.
}
\keyword{datasets}
