% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onsitehabitat.R
\name{pullA1baseline}
\alias{pullA1baseline}
\title{pull onsite baseline data for summary}
\usage{
pullA1baseline(metric)
}
\arguments{
\item{metric}{feas metric}
}
\value{
a list, containing dataframes pulled from the baseline sheet of the metric
}
\description{
pull onsite baseline data for summary
}
\examples{
\dontrun{baselinedata<-pullA1baseline(metric)}
}
