% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onsitehabitat.R
\name{pullA1loss}
\alias{pullA1loss}
\title{pull A1 loss}
\usage{
pullA1loss(metric)
}
\arguments{
\item{metric}{feas metric}
}
\value{
a list, containing dataframes pulled from the baseline sheet of the metric
}
\description{
pull A1 loss
}
\examples{
\dontrun{lostdata<-pullA1loss(metric)}
}
