% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onsitehabitat.R
\name{pullA1retain}
\alias{pullA1retain}
\title{pull A1 retain}
\usage{
pullA1retain(metric)
}
\arguments{
\item{metric}{feas metric}
}
\value{
a list, containing dataframes pulled from the baseline sheet of the metric
}
\description{
pull A1 retain
}
\examples{
\dontrun{retaindata<-pullA1retain(metric)}
}
