% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onsitehabitat.R
\name{pullA2}
\alias{pullA2}
\title{pull onsite habitats to be created}
\usage{
pullA2(metric)
}
\arguments{
\item{metric}{feas metric}
}
\value{
a list, containing data about habitats to be created
}
\description{
pull onsite habitats to be created
}
\examples{
\dontrun{creationdata<-pullA2(metric)}
}
