% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onsite_habitat_metricfunctions.R
\name{pullonsitehabitatbaseline}
\alias{pullonsitehabitatbaseline}
\title{On site habitat baseline}
\usage{
pullonsitehabitatbaseline(metric)
}
\arguments{
\item{metric}{file path to metric}
}
\value{
list with data.frame of baseline data, total area nd total baseline units
}
\description{
On site habitat baseline
}
\examples{
\dontrun{baselinedata<-pullonsitehabitatbaseline(metric)}
}
