% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onsite_habitat_metricfunctions.R
\name{pullonsitehabitatcreation}
\alias{pullonsitehabitatcreation}
\title{pull onsite habitats to be created}
\usage{
pullonsitehabitatcreation(metric)
}
\arguments{
\item{metric}{feas metric}
}
\value{
a list, containing data about habitats to be created
}
\description{
pull onsite habitats to be created
}
\examples{
\dontrun{creationdata<-pullonsitehabitatcreation(metric)}
}
