% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onsite_habitat_metricfunctions.R
\name{pullonsitehabitatenhance}
\alias{pullonsitehabitatenhance}
\title{On-site habitat enhancement summary data}
\usage{
pullonsitehabitatenhance(metric)
}
\arguments{
\item{metric}{A BNG metric}
}
\value{
A list with a summary data.frame, an numeric for total habitat area
enhanced and a numeric for total BNG units enhanced
}
\description{
On-site habitat enhancement summary data
}
\examples{
\dontrun{enhancedata <- pullonsitehabitatenhance(metric)}
}
