% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onsite_habitat_metricfunctions.R
\name{pullonsitehabitatenhancement}
\alias{pullonsitehabitatenhancement}
\title{pull onsite habitats to be enhanced}
\usage{
pullonsitehabitatenhancement(metric)
}
\arguments{
\item{metric}{feas metric}
}
\value{
a list, containing 3 elements: BroadEnhance, the broad habitat type, e.g 'Grassland',
EnhanceHabitatType, the detailed habitat type, e.g. 'Other neutral grassland', and then
EnhanceList, a list of strings summarising the habitat types proposed for enhancement on the site.
}
\description{
pull onsite habitats to be enhanced
}
\examples{
\dontrun{enhancedata<-pullonsitehabitatenhancement(metric)}
}
