% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onsite_habitat_metricfunctions.R
\name{pullonsitehabitatloss}
\alias{pullonsitehabitatloss}
\title{pullhabitatloss}
\usage{
pullonsitehabitatloss(metric)
}
\arguments{
\item{metric}{metric}
}
\value{
loss df
}
\description{
pullhabitatloss
}
\examples{
\dontrun{habitatlossdata<-pullonsitehabitatloss(metric)}
}
