% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onsite_habitat_metricfunctions.R
\name{pullonsitehabitatretain}
\alias{pullonsitehabitatretain}
\title{pullhabitatretain}
\usage{
pullonsitehabitatretain(metric)
}
\arguments{
\item{metric}{metric}
}
\value{
retain df
}
\description{
pullhabitatretain
}
\examples{
\dontrun{retaindata<-pullonsitehabitatretain(metric)}
}
