% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onsite_tradingsummary_metricfunctions.R
\name{pullonsitehabitatsumdata}
\alias{pullonsitehabitatsumdata}
\title{pull metadata needed, including site size, net units gained, net percentage gain, etc.}
\usage{
pullonsitehabitatsumdata(metric)
}
\arguments{
\item{metric}{feas metric}
}
\value{
sumdata, a list containing summarising data about the site, including potential net BNG units
gained, BNGUnits, and the net bng percentage increase, BNGPercentage, and Hectares, the size of the site.
}
\description{
pull metadata needed, including site size, net units gained, net percentage gain, etc.
}
\examples{
\dontrun{habitatsumdata<-pullonsitehabitatsumdata(metric)}
}
