% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onsite_hedge_metricfunctions.R
\name{pullonsitehedgebaseline}
\alias{pullonsitehedgebaseline}
\title{pulls on site hedge baseline data}
\usage{
pullonsitehedgebaseline(metric)
}
\arguments{
\item{metric}{metric}
}
\value{
hedgebaselinedata
}
\description{
pulls on site hedge baseline data
}
\examples{
\dontrun{hedgebaselinedata<-pullonsitehedgebaseline(metric)}
}
