% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onsite_hedge_metricfunctions.R
\name{pullonsitehedgecreation}
\alias{pullonsitehedgecreation}
\title{pull created hedge data}
\usage{
pullonsitehedgecreation(metric)
}
\arguments{
\item{metric}{metric}
}
\value{
hedgecreationdata
}
\description{
pull created hedge data
}
\examples{
\dontrun{hedgecreation<-pullonsitehedgecreation(metric)}
}
