% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onsite_hedge_metricfunctions.R
\name{pullonsitehedgeenhancement}
\alias{pullonsitehedgeenhancement}
\title{pull hedge enhance}
\usage{
pullonsitehedgeenhancement(metric)
}
\arguments{
\item{metric}{metric}
}
\value{
ehnacne hedge data
}
\description{
pull hedge enhance
}
\examples{
\dontrun{hedgeenhancementdata<-pullonsitehedgeenhancement(metric)}
}
