% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onsite_hedge_metricfunctions.R
\name{pullonsitehedgeloss}
\alias{pullonsitehedgeloss}
\title{pullhedgeloss}
\usage{
pullonsitehedgeloss(metric)
}
\arguments{
\item{metric}{metric}
}
\value{
lost hedge data
}
\description{
pullhedgeloss
}
\examples{
\dontrun{hedgelossdata<-pullonsitehedgeloss(metric)}
}
