% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onsite_hedge_metricfunctions.R
\name{pullonsitehedgeretain}
\alias{pullonsitehedgeretain}
\title{pullhedgeretain}
\usage{
pullonsitehedgeretain(metric)
}
\arguments{
\item{metric}{metric}
}
\value{
retain hedge data
}
\description{
pullhedgeretain
}
\examples{
\dontrun{hedgeretaindata<-pullonsitehedgeretain(metric)}
}
