% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onsite_net_metricfunctions.R
\name{pullonsitenetdata}
\alias{pullonsitenetdata}
\title{pull data regarding the net unit gains and net percentage gains, as well as if loss present}
\usage{
pullonsitenetdata(metric)
}
\arguments{
\item{metric}{feas metric}
}
\value{
netdata, a list containing summarising data about the site, including potential net BNG units
gained, BNGUnits, and the net bng percentage increase, BNGPercentage, and Hectares, the size of the site.
}
\description{
pull data regarding the net unit gains and net percentage gains, as well as if loss present
}
\examples{
netdata <- pullonsitenetdata(metric = system.file("extdata", "OnSiteBoth.xlsx", package = "metricpulling"))
}
