% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onsite_water_metricfunctions.R
\name{pullonsitewaterbaseline}
\alias{pullonsitewaterbaseline}
\title{pulls on site water baseline data}
\usage{
pullonsitewaterbaseline(metric)
}
\arguments{
\item{metric}{metric}
}
\value{
waterbaselinedata
}
\description{
pulls on site water baseline data
}
\examples{
\dontrun{waterbaselinedata<-pullonsitewaterbaseline(metric)}
}
