% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onsite_water_metricfunctions.R
\name{pullonsitewatercreation}
\alias{pullonsitewatercreation}
\title{pull created water data}
\usage{
pullonsitewatercreation(metric)
}
\arguments{
\item{metric}{metric}
}
\value{
watercreationdata
}
\description{
pull created water data
}
\examples{
\dontrun{watercreation<-pullonsitewatercreation(metric)}
}
