% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onsite_water_metricfunctions.R
\name{pullonsitewaterenhancement}
\alias{pullonsitewaterenhancement}
\title{pull water enhance}
\usage{
pullonsitewaterenhancement(metric)
}
\arguments{
\item{metric}{metric}
}
\value{
ehnacne water data
}
\description{
pull water enhance
}
\examples{
\dontrun{waterenhancementdata<-pullonsitewaterenhancement(metric)}
}
