% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onsite_water_metricfunctions.R
\name{pullonsitewaterloss}
\alias{pullonsitewaterloss}
\title{pull water loss}
\usage{
pullonsitewaterloss(metric)
}
\arguments{
\item{metric}{metric}
}
\value{
water loss data
}
\description{
pull water loss
}
\examples{
\dontrun{waterlossdata<-pullonsitewaterloss(metric)}
}
