% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onsite_water_metricfunctions.R
\name{pullonsitewaterretain}
\alias{pullonsitewaterretain}
\title{pull water retain}
\usage{
pullonsitewaterretain(metric)
}
\arguments{
\item{metric}{metric}
}
\value{
water retain data
}
\description{
pull water retain
}
\examples{
\dontrun{waterretaindata<-pullonsitewaterretain(metric)}
}
