% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rendersurveymethodology.R
\name{rendersurveymethodology}
\alias{rendersurveymethodology}
\title{render survey dynamically}
\usage{
rendersurveymethodology(survey, sectionCounter)
}
\arguments{
\item{survey}{survey methodology to input}

\item{sectionCounter}{counter for title number}
}
\value{
section counter, to be passed to next dynamically rendered subsection
}
\description{
render survey dynamically
}
\examples{
\dontrun{sectioncounter<-rendersurveymethodology("Grass", sectionCounter)}
}
