% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearpilbusummary.R
\name{waterlbupicreate}
\alias{waterlbupicreate}
\title{waterlbupicreate}
\usage{
waterlbupicreate(watercreationdata, sectioncounter)
}
\arguments{
\item{watercreationdata}{water creation df}

\item{sectioncounter}{counter passed to following dynamic area based post intervention sections}
}
\value{
sectioncounter, passed to future dynamically rendered sections under post intervention
}
\description{
waterlbupicreate
}
\examples{
\dontrun{sectioncounter<-waterlbupicreate(watercreationdata$watercreationdata, sectioncounter)}
}
