% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearpilbusummary.R
\name{waterlbupienhance}
\alias{waterlbupienhance}
\title{linear enhance}
\usage{
waterlbupienhance(waterenhancementdata, sectioncounter)
}
\arguments{
\item{waterenhancementdata}{waterenhancementdata df}

\item{sectioncounter}{counter to pass}
}
\value{
counter to pass
}
\description{
linear enhance
}
\examples{
\dontrun{sectioncounter<-waterlbupienhance(waterenhancedata$waterenhancementdata, sectioncounter)}
}
