% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearpilbusummary.R
\name{waterlbupilost}
\alias{waterlbupilost}
\title{waterlbupilost}
\usage{
waterlbupilost(waterlostdata, sectioncounter)
}
\arguments{
\item{waterlostdata}{water loss df}

\item{sectioncounter}{counter passed to following dynamic area based post intervention sections}
}
\value{
sectioncounter, passed to future dynamically rendered sections under post intervention
}
\description{
waterlbupilost
}
\examples{
\dontrun{sectioncounter <- waterlbupilost(waterlossdata$waterlostdata, sectioncounter)}
}
