% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearpilbusummary.R
\name{waterlbupiretain}
\alias{waterlbupiretain}
\title{waterlbupiretain}
\usage{
waterlbupiretain(waterretaindata, sectioncounter)
}
\arguments{
\item{waterretaindata}{water retain df}

\item{sectioncounter}{counter passed to following dynamic area based post intervention sections}
}
\value{
sectioncounter, passed to future dynamically rendered sections under post intervention
}
\description{
waterlbupiretain
}
\examples{
#' @examples \dontrun{sectioncounter<-waterlbupiretain(waterretaindata$waterretaindata, sectioncounter)}
}
