% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterlbusummary.R
\name{waterlbusummary}
\alias{waterlbusummary}
\title{water lbu summary}
\usage{
waterlbusummary(waterbaselinedata, sectioncounter)
}
\arguments{
\item{waterbaselinedata}{data for baseline watercourses}

\item{sectioncounter}{counter for dynamic rendering}
}
\value{
rendered section
}
\description{
water lbu summary
}
\examples{
\dontrun{waterlbusummary(waterbaselinedata, sectioncounter)}
}
