% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data_catches.R
\name{clean_C1_dataset}
\alias{clean_C1_dataset}
\title{cleans C1 water metric data}
\usage{
clean_C1_dataset(metric)
}
\arguments{
\item{metric}{feas metric}
}
\value{
df, a cleaned version of the input dataset - to be used as an argument in actual metric pulling function
}
\description{
cleans C1 water metric data
}
\examples{
\dontrun{checked_dataset<-clean_C1_dataset(metric)}  
}
