% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_check.R
\name{metric_check}
\alias{metric_check}
\title{metric_check}
\usage{
metric_check(metric)
}
\arguments{
\item{metric}{metric}
}
\value{
text with list of where data is missing in metric
}
\description{
metric_check
}
\examples{
\dontrun{
a <- metric_check("data-raw/OnSiteHedgeEnhance.xlsx") 
b <- metric_check("data-raw/OnSiteHedgeEnhanceMissing.xlsx") # intentionally bad data
c <- metric_check("data-raw/OnSiteHedgeEnhance2.xlsx") # metric used in testing qmd}
}
