% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onsite_habitat_metricfunctions.R
\name{pullonsitehabitatbaseline}
\alias{pullonsitehabitatbaseline}
\title{pull onsite baseline data for summary}
\usage{
pullonsitehabitatbaseline(metric)
}
\arguments{
\item{metric}{feas metric}
}
\value{
a list, containing dataframes pulled from the baseline sheet of the metric
}
\description{
pull onsite baseline data for summary
}
\examples{
\dontrun{baselinedata<-pullonsitehabitatbaseline(metric)}
}
