% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onsite_habitat_metricfunctions.R
\name{pullonsitehabitatcreation}
\alias{pullonsitehabitatcreation}
\title{On site habitat created function for testing}
\usage{
pullonsitehabitatcreation(metric)
}
\arguments{
\item{metric}{file path to metric}
}
\value{
list with data.frame of habitats created, total area of habitats created, and total units created
}
\description{
On site habitat created function for testing
}
\examples{
\dontrun{pullonsitehabitatcreation(metric)}
}
