% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onsite_habitat_metricfunctions.R
\name{pullonsitehabitatloss}
\alias{pullonsitehabitatloss}
\title{On site habitat lost function}
\usage{
pullonsitehabitatloss(metric)
}
\arguments{
\item{metric}{file path to metric}
}
\value{
list with data.frame of habitats lost, total area lost, and total units lost
}
\description{
On site habitat lost function
}
\examples{
\dontrun{pullonsitehabitatloss(metric)}
}
