% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onsite_habitat_metricfunctions.R
\name{pullonsitehabitatretain}
\alias{pullonsitehabitatretain}
\title{On site habitats retained}
\usage{
pullonsitehabitatretain(metric)
}
\arguments{
\item{metric}{file path to metric}
}
\value{
list with data.frame of retained habitats, total retained area, total retained units
}
\description{
On site habitats retained
}
\examples{
\dontrun{pullonsitehabitatretain(metric)}
}
